refactor: annotate squad/
Takes squad.history, squad.player, squad.squad and squad.state off the mypy
migration list. squad.py was the second-largest entry, with 22 of its 25
functions unannotated.

A Squad holds a CandidatePlayer or a DummyPlayer, so that union is now named
(SquadPlayer) rather than restated - the one existing annotation of it, on
optimize_subs' player_dict, said CandidatePlayer and was wrong.

Writing the types down turned up five real problems:

- DummyPlayer.player_id was a uuid *string* while a real player's is an int, so
  a squad's player ids were int | str, and a dummy left in an optimised squad
  wrote that string into TransferSuggestion.player_id, an integer column. Dummy
  ids are now negative ints: same type as a real id, still unique, still
  obviously placeholders.
- sub_position is None until the lineup has been optimized, and four places
  sorted or indexed on it directly, failing inside sorted() with a TypeError
  about None. They now go through squad.player.bench_position, which says what
  went wrong.
- optimize_subs passed best_formation to apply_formation without checking it
  had been set.
- get_bank returned float for a value in tenths of a million that is assigned
  straight to Squad.budget, an int.
- get_playerscores_for_player_gameweek asked for a Player but only ever used
  its id, and both callers hold a squad player rather than an ORM row. It takes
  the id now.
- fill_initial_squad and update_squad resolved fpl_team_id themselves without
  checking the result - exactly what require_fpl_team_id exists for.

+++ src/airsenal/squad/state.py
@@ -20,7 +20,7 @@ def get_bank(
     gameweek: int | None = None,
     season: str = CURRENT_SEASON,
     fetcher: FPLDataFetcher | None = None,
-) -> float:
+) -> int:
     """
     Find out how much this FPL team had in the bank before the specified gameweek.
     If gameweek is not provided, give the most recent value.
@@ -51,9 +51,9 @@ def get_bank(
         if gameweek and isinstance(gameweek, int):
             for gw in data["current"]:
                 if gw["event"] == gameweek - 1:  # value after previous gameweek
-                    return gw["bank"]
+                    return int(gw["bank"])
         # otherwise, return the most recent value
-        return data["current"][-1]["bank"]
+        return int(data["current"][-1]["bank"])
 
 
 def get_entry_start_gameweek(
